Use delayedWrite instead of request
That way all api requests go through the queue. If we don't have them go
through the queue a bunch of them can fire off in parallel, get 407, and
then be replayed.

By going through the queue we only replay the first api request that
fails with 407, and after we authenticate we fire off all the other
requests