DatePicker: mWeb show datepicker on focus
We're using ExpensiInput which override some styles and render the input differently
There's no dropdown arrow.
When we focus the empty datepicker field it's only highlighted in blue. The user
needs to tap again in order to see the native datepicker
To avoid this confusion we're showing the datepicker immediately