Move form validation errors to component state.
I don't see a reason to keep these form errors in Onyx: there is no
use in persisting them outside of the life of the form.
Onyx doesn't support a way of merging a value using a callback in the
same way as setState. The AddressSearch component can make us clear
an error of several fields at the same time. Without the callback mode
to set a state, clearing the error of multiple fields individually one
after the other causes the errors not to be properly cleared (subsequent
calls keep using the old state).