Update setting event options
Updated the way event data is set for `Act` instance, since `Act` seems to not be consistent in how it expects the data to be provided based on the type of event. This change seems to cover all possibilities. Plus renamed some variables to be compliant with JS convention

See: https://github.com/Expensify/App/issues/13604