Imperfect version of redirect prototype for apple
Two problems:

- No good way to send the token over with this prototype. The "." in the
  tokens appears to break URL parsing. Shouldn't be sending Apple token
in URL anyways, but I can't generate short tokens, and the other way to
send token over would be like, cookies, and that would only work from
POST->web app, wouldn't work for desktop
- This flow technically needs an additional route; one to receive the
  POST redirect to web app, and then another to receive the token from
desktop. But not bothering with this because it's kind of up to how
backend is implemented (do we get a short-lived token, does POST
redirect to desktop directly or let web decide).