Fix: Chat - Link in end of line displays tooltip over text and not on link
Add a new prop to Tooltip called shouldUseMultilinePositioning. This
enables a new algorithm for finding the bounding box target for the
tooltip. In the case where the target has wrapped onto multiple lines
(e.g. it is a link in a chat window) then the link has multiple bounding
boxes, and we want to show the tooltip against the one that the user is
hovering over.

As part of this, extend Hoverable to pass the Event to its onHoverIn /
onHoverOut callbacks.

Enable this new algorithm from BaseAnchorForCommentsOnly, which is the
base class for links that show up in chat.