Fix: Chat - Link in end of line displays tooltip over text and not on link
Changes from PR review:
  o  Remove the shouldUseMultilinePositioning prop, and instead make the
new algorithm the one to use in all situations.
  o  Refactor chooseBoundingBox to avoid a recursive call.
  o  Shortcut chooseBoundingBox in the case where the element only
has one bounding box.
  o  Revert the changes to Hoverable that pass the MouseEvent through
onHoverIn / onHoverOut, and instead add direct pass-throughs for the
onMouseEnter / onMouseLeave events.
  o  Change the algorithm so that the onHoverIn event handler only
records the mouse position, and rely on BoundsObserver being enabled
as a side-effect of the isVisible change to move the tooltip.