Create "Hidden" folder in helpsite repo to house delisted articles
Per https://expensify.slack.com/archives/C02QSAC6BJ8/p1714064970851609?thread_ts=1713874136.782669&cid=C02QSAC6BJ8 I have been tasked with creating the Hidden folder in the helpsite. 

We intend to use this folder to host articles that we do not want to indexed or searchable on the help site. Customers should not be able to access these.