wip, current problem:
The current problem is with the max reports number. I currently only splice in the order function. but in the filter function we filter the personal details based on the recent reports. if the recent reports aren't limited in the filter function then then we will have fewer personal details (as more reports have been loaded).

however i can't limit the recent reports in filter yet as they aren't ordered yet…

we might need to split up getOptions for reports and pD actually