Create personal-cards.md
Tracking GH: https://github.com/Expensify/Expensify/issues/462432

Creating a personal card page in NewDot help site. This feature isn't available yet so this will be a placeholder. The intention of the article is to provide customers with a self-service method to find out more about personal cards in NewDot.